requirements.txt
streamlit==1.32.0
pandas==2.2.1
openai==1.12.0
python-dotenv==1.0.1
openpyxl==3.1.2